Add script to update firmware and ability to modify the environment
The script is not working correctly at the moment. There are issues
with NOR access that cause timeout and halts write. The write has to be
attempted again after that. The second issue is that sometimes cmp
reports difference even when the content is the same.

Alongside with this this also adds configuration to access the U-Boot
environment. This seems to work correctly, although there are possible
ways to improve it.