nixos/hostapd: change back to use upstream's services.hostapd
The revert to upstream is to allow easier merge of the module to the
upstream.

The primary other change here is that every interface now has its own
service.  The reasoning is to just to allow easier control of separate
interfaces.

The option vht_oper_centr_freq_seg0_idx is now optional the same way as
vht_oper_centr_freq_seg1_idx is. The seq0 should not be in the
configuration file when channel width is 40MHz or bellow. The seq1
should not be in the configuraiton file when channel width is 80MHz or
less. The result of having them is hostapd unable to calculate valid
frequency.

The library functions are now also rather in form of hostapd
configuration expansion.