nixos: rework tarball and cross compilation packages
The dedicated cross compilation packages are not essentially required as
we can easilly just select different host architecture to not cross
compile and thus this provides rather cross compilation in default.

This also reworks tarball to just do tarball. The install stuff is
included on the side with another module.