nixos: rework the initial configuration
It would be goot to provide images but right now there is no module to
create btrfs file system image and this we can't easilly create that
image. For time being the tar is the best option still.

The result of this is less hacky initial extlinux file (generated by
proper tool) and correct registration on the first boot. The
installation has to be still performed the same way as previously.