pkgs: reduce size of the kernel
This reduces size of the kernel. We need to perform compilation due to
added patches anyway so we are loosing advantage of having kernel that
can be deployed to multiple targets. Thus it makes sense for us to
rather disable as many of the as possible to reduce size and speed up
the compilation.